Add message delete event handler
