+++ events/message.py
@@ -1,6 +1,9 @@
 import inspect
 import sys
 
+import discord
+from helpers.embed_builder import EmbedBuilder
+
 from events.base import EventHandler
 
 
@@ -30,6 +33,36 @@ class MessageEvent(EventHandler):
                 await command_handler(self.client).execute(message, command=cmd, args=command, storage=self.client.storage, instance=self.client)
             else:
                 await message.channel.send("**Unknown command:** {}".format(cmd))
+
+
+class MessageDeleteEvent(EventHandler):
+    def __init__(self, client_instance):
+        self.client = client_instance
+        self.event = "on_message_delete"
+    
+    async def handle(self, *args, **kwargs):
+        # Get the message from the args
+        message = args[0]
+
+        # Ignore deletes of bot messages or messages from ourselves
+        if message.author == self.client.user or message.author.bot:
+            return
+        # Build an embed that will log the deleted message
+        embed_builder = EmbedBuilder(event="delete")
+        await embed_builder.add_field(name="**Channel**", value=f"`#{message.channel.name}`")
+        await embed_builder.add_field(name="**Author**", value=f"`{message.author.name}`")
+        await embed_builder.add_field(name="**Message**", value=f"`{message.content}`")
+        await embed_builder.add_field(name="**Created at**", value=f"`{message.created_at}`")
+        embed = await embed_builder.get_embed()
+        
+        # Message the log channel the embed of the deleted message
+        guild_id = str(message.guild.id)
+        log_channel_id = int(self.client.storage.settings["guilds"][guild_id]["log_channel_id"])
+        log_channel = discord.utils.get(message.guild.text_channels, id=log_channel_id)
+        if log_channel is not None:
+            await log_channel.send(embed=embed)
+        else:
+            print("No log channel found with that ID")
 
 
 # Collects a list of classes in the file
